server.js get and post routes are moved into the controllers
