+++ controllers/api/userRoutes.js
@@ -1,43 +1,52 @@
 const router = require('express').Router();
 const { User } = require('../../model');
 
-// Portion of controller for logging in
-router.post('/login', async (req, res) => {
+app.post('/login', async (req, res) => {
     try {
-        // Email that was input is searched within the model's user information
-        const userData = await User.findOne({ where: { email: req.body.email } });
-
-        // If no email is found, display message below
-        if (!userData) {
-            res
-                .status(400)
-                .json({ message: 'Incorrect email, please try again' });
-            return;
+        const { loginIdentifier, loginPassword } = req.body;
+    
+        const user = User.find(user => user.email === loginIdentifier || user.username === loginIdentifier);
+        if (!user) {
+        return res.status(404).json({ message: 'User not found' });
         }
-
-        // Validate a user's password matches the email it is connected to
-        const validPassword = await userData.checkPassword(req.body.password);
-
-        // If not valid, display message telling the user it is incorrect
-        if (!validPassword) {
-            res
-                .status(400)
-                .json({ message: 'Incorrect password, please try again' });
-            return;
+    
+        if (User.password !== loginPassword) {
+        return res.status(401).json({ message: 'Incorrect password' });
         }
+    
+        res.json({ message: 'Login successful' });
 
-        // Request session information then responds with successful login
         req.session.save(() => {
             req.session.user_id = userData.id;
             req.session.logged_in = true;
 
             res.json({ user: userData, message: 'Successfully logged in!' });
         });
+
     } catch (err) {
         console.error(err);
         res.status(400).json(err);
     }
-});
+  });
+
+app.post('/signup', (req, res) => {
+    const { email, username, password, animalChoice} = req.body;
+  
+    const userExists = User.some(user => user.email === email || user.username === username);
+    if (userExists) {
+      return res.status(400).json({ message: 'User already exists '});
+    }
+  
+    const newUser = { email, username, password, animalChoice };
+    user.push(newUser);
+  
+    res.status(201).json({ message: 'User registered successfully' });
+  });
+
+app.get('/', (req, res) => {
+    // THIS NEEDS TO BE DONE
+    // USE res.render or res.sendFile
+  });
 
 // If a user logs out, it destroys the session and logs the user out. If unable to, throws a 404 error
 router.post('/logout', async (req, res) => {
