--- controllers/homeRoutes.js
@@ -1,13 +1,0 @@
-// If not logged in, directs user to login before viewing homepage.
-// If logged in, homepage is displayed
-router.get('/login', (req, res) => {
-    if (req.session.logged_in) {
-      res.redirect('/');
-      return;
-    }
-  
-    res.render('login');
-  });
-  
-  module.exports = router;
-
