+++ package.json
@@ -18,7 +18,6 @@
     "express": "^4.16.3",
     "express-handlebars": "^7.1.2",
     "express-session": "^1.17.3",
-    "multer": "^1.4.5-lts.1",
     "mysql2": "^2.2.5",
     "sequelize": "^6.3.5"
   }
